fix windows int type error

+++ pypeit/bitmask.py
@@ -351,12 +351,12 @@ class BitMask:
 
         out = value | (1 << self.bits[_flag[0]])
         if len(_flag) == 1:
-            return out
+            return out.astype(value.dtype)
 
         nn = len(_flag)
         for i in range(1,nn):
             out |= (1 << self.bits[_flag[i]])
-        return out
+        return out.astype(value.dtype)
 
     def turn_off(self, value, flag):
         """
